test: Add test and snapshot for emails
