+++ package.json
@@ -224,14 +224,17 @@
     "moduleNameMapper": {
       "\\.(png|gif|jpe?g|svg)$": "<rootDir>/test/__mocks__/fileMock.js",
       "styl$": "identity-obj-proxy",
-      "\\.css$": "identity-obj-proxy",
-      "hbs$": "identity-obj-proxy",
-      "webapp$": "identity-obj-proxy"
+      "webapp$": "identity-obj-proxy",
+      "!!raw-loader!(.*)": "$1"
     },
     "testPathIgnorePatterns": [
       "node_modules",
       "src/targets/mobile/"
     ],
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest",
+      "\\.(hbs|css)$": "<rootDir>/test/readFile.js"
+    },
     "transformIgnorePatterns": [
       "node_modules/(?!(cozy-ui|cozy-client))"
     ],
