+++ test/readFile.js
@@ -1,0 +1,5 @@
+module.exports = {
+  process: function(fileContent) {
+    return `module.exports = \`${fileContent}\``
+  }
+}
